refactor: add price domain
