+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -3,11 +3,13 @@ package kitchenpos.product.domain;
 import java.math.BigDecimal;
 import java.util.Objects;
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import kitchenpos.common.domain.Price;
 
 @Entity
 @Table(name = "product")
@@ -16,8 +18,9 @@ public class Product {
     private Long id;
     @Column(nullable = false)
     private String name;
-    @Column(nullable = false)
-    private BigDecimal price;
+
+    @Embedded
+    private Price price;
 
     protected Product() {
     }
@@ -29,7 +32,7 @@ public class Product {
     public Product(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
-        this.price = price;
+        this.price = new Price(price);
 
         validatePrice();
     }
@@ -42,21 +45,13 @@ public class Product {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Price getPrice() {
         return price;
     }
 
     private void validatePrice() {
-        if (isPriceNull() || isPriceLessThanZero()) {
+        if (price.isNull() || price.isLessThan(BigDecimal.ZERO)) {
             throw new IllegalArgumentException();
         }
-    }
-
-    private boolean isPriceLessThanZero() {
-        return price.compareTo(BigDecimal.ZERO) < 0;
-    }
-
-    private boolean isPriceNull() {
-        return Objects.isNull(price);
     }
 }
